+++ docs/make.jl
@@ -2,6 +2,9 @@ include(joinpath(@__DIR__, "activate.jl"))
 
 using AtmosphericDispersion
 using Documenter
+using DocumenterCitations
+
+bib = CitationBibliography(joinpath(@__DIR__, "src", "refs.bib"); style = :authoryear)
 
 DocMeta.setdocmeta!(
     AtmosphericDispersion,
@@ -14,6 +17,7 @@ makedocs(;
     modules = [AtmosphericDispersion],
     authors = "Paul-Adrian Gogîță",
     sitename = "AtmosphericDispersion.jl",
+    plugins = [bib],
     format = Documenter.HTML(;
         canonical = "https://PaulGoG.github.io/AtmosphericDispersion.jl",
         edit_link = "main",
@@ -30,6 +34,7 @@ makedocs(;
             "Dilution, depletion and deposition" => "api/fields.md",
             "Configuration" => "api/configuration.md",
         ],
+        "References" => "references.md",
     ],
 )